requirements.txt
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.11.29
chardet==3.0.4
Django==2.1.1
django-ckeditor==5.6.1
django-js-asset==1.1.0
idna==2.8
lxml==4.2.5
Pillow==5.2.0
pkg-resources==0.0.0
pytz==2018.5
requests==2.21.0
urllib3==1.24.1
    